backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
notion-client
pydantic
python-dotenv
langsmith
